iio-utils: remove warning from VS2019_Win64
scan.c(167,33): warning C4267: '=': conversion from 'size_t' to 'unsigned int', possible loss of data
iio_writedev.c(500,65): warning C4244: '=': conversion from 'uint64_t' to 'unsigned int', possible loss of data
iio_readdev.c(455,65): warning C4244: '=': conversion from 'uint64_t' to 'unsigned int', possible loss of data

so just cast strlen to a (unsigned int); and use uint64_t as default

+++ scan.c
@@ -164,7 +164,7 @@ struct iio_scan_context * iio_create_scan_context(
 
 	if (backend) {
 		/* Replace the colon separator with a comma. */
-		len = strlen(ctx->backendopts);
+		len = (unsigned int)strlen(ctx->backendopts);
 		for (i = 0; i < len; i++)
 			if (ctx->backendopts[i] == ':')
 				ctx->backendopts[i] = ',';
